Extract `localStorage` data providing with custom hook
Fingers crossed that this won't break `FirebaseUserDataProvider` :pray:
